fix: unknown types should not be accepted by default
Originally, the validation on Conventional Commit types was only
working correctly in case the user would provide a list of
additional types.

This would result in the following commit to be accepted:
```
feet: feed it
```

Worst case, this would lead to an absence of a SemVer bump
in case automation relies on the types.