feat: add support for Conventional Commit spec #06
Item 6. describes:
> A longer commit body MAY be provided after the short description,
> providing additional contextual information about the code changes.
> The body MUST begin one blank line after the description.

Unfortunately, in case of a commit message such as:

```
feat: this is wrong
because we have no space here
```

we would ignore the error and report incorrect errors to the end-user
(as it could not parse the subject for types, scopes, etc).

This commit introduces a rule specifically to cover subjects which
cover multiple lines.